Fix bfloat16 → float32 conversion in feature extraction
Pi0.5 outputs bfloat16 hidden states. numpy doesn't natively support
bfloat16, so np.asarray() produces an object array that sklearn rejects.
Explicitly cast to float32.

+++ src/openpi/contact_mpc/features/extractor.py
@@ -30,7 +30,7 @@ def extract_features_from_observation(
         Pooled hidden states as numpy array of shape [batch_size, hidden_dim].
     """
     pooled = model.extract_vlm_features(observation)
-    return np.asarray(pooled)
+    return np.asarray(pooled, dtype=np.float32)
 
 
 def extract_features_from_dict(
